requirements.txt
pandas
request
streamlit